backend/requirements.txt
# backend/requirements.txt
Flask==2.3.2
Flask-SQLAlchemy==3.0.3
Flask-Migrate==4.0.4
python-dotenv==1.0.1
requests==2.31.0
pytest==7.3.1